LMREADY actually takes a `u_long' (really, a pointer).
LKM's work on sparc64 now, and might on the alpha.  note that this changes
the value of `LMREADY' for LP64, however, it is currently broken anyway..

you will need to rebuild both the kernel and modload(1) (with the changed
<sys/lkm.h>) for this change to be effective.

+++ sys/sys/lkm.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: lkm.h,v 1.19 2000/12/08 19:42:12 jdolecek Exp $	*/
+/*	$NetBSD: lkm.h,v 1.20 2001/02/25 17:44:15 mrg Exp $	*/
 
 /*
  * Header file used by loadable kernel modules and loadable kernel module
@@ -316,7 +316,7 @@ extern int	lkmdispatch __P((struct lkm_table *, int));
 #define	LMRESERV_O	_IOWR('K', 0, struct lmc_oresrv)
 #define	LMLOADBUF	_IOW('K', 1, struct lmc_loadbuf)
 #define	LMUNRESRV	_IO('K', 2)
-#define	LMREADY		_IOW('K', 3, int)
+#define	LMREADY		_IOW('K', 3, u_long)
 #define	LMRESERV	_IOWR('K', 4, struct lmc_resrv)
 
 #define	LMLOAD		_IOW('K', 9, struct lmc_load)
